Move tile endpoints out of views
